+++ tests/bootstrap.php
@@ -41,7 +41,6 @@ define('CACHE', TMP . 'cache' . DS);
 define('LOGS', TMP . 'log' . DS);
 define('SESSIONS', TMP . 'sessions' . DS);
 define('UPLOADED', WWW_ROOT . 'files' . DS);
-define('LOGIN_RECORDS', TMP . 'login' . DS);
 
 foreach ([
     TMP . 'tests',