+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.30 branch
 ### 2.30.7-RC4
+* added `last_logins` field to user table. This field is used instead of files.
+    So the `LoginRecorderComponent` no longer uses files, but the database;
 * updated for PHP 8.1.
 
 ### 2.30.6-RC3